Созданы модули settings, utils, create-advertisement и подключены в файл main.js
